requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.5
django-crispy-forms==2.0
django-tinymce==3.6.1
sqlparse==0.4.4
tzdata==2023.3
